Fix Lax-Friedrichs
Old Lax-Friedrichs had issues with the signs.
Had to 'hardcode' upwind in previous versions to obtain optimal output error of 2p+1.
Since Lax-Friedrichs and upwind are the same for linear advection, should obtain the same.

Note that for 2D, p=3, still get suboptimal output error of 2p instead of 2p+1. Not sure why only that case.