Small fix for major bug visit face twice MPI
This problem did not occur when I used a lower number of processors.

The bug occurs when visiting the face of a cell that has a ghost cell neighbor.
In that scenario, it used to be the cell with the lower index who was responsible to perform the work on the current face.
However, when a distributed mesh is stored, that index is not necessarily consistent across processors.
Therefore, sometimes, both processors agreed on which cell did the work, other times, the work done on that face was performed twice.