Working d2R FD vs AD.
Initializing the FD with the interpolated manufactured solution caused
some issue at the boundary condition where the value was exactly the
manufactured solution. As a result, it was at the critical point where
the switch between inflow/outflow happens. In the numerical flux, this
results in differentiating the "max" function at the switching point,
which is of course non-differentiable.

Fixed by changing the inital condition.