Update regression test.
The GridRefinement pull request used the parameter file to specify the constants within the manufactured solution.
This resulted in small differences in the constants being used and lead to 1e-6 differences in the Jacobians generated.

This has no impact on the actual code, so we simply update our regression tests.