Better non-physical value handling and adding parmeters object to Physics (#214)
A recent change prints warning messages whenever a negative
density/pressure is detected in Euler. This is helpful for some tests,
but results in very long log files for others (for instance, where an
iterative solver sometimes fails but the test is built to handle this).
Therefore, it is useful to add a parameter which will give the user more
control over behaviour in these cases.

A function has been added to PhysicsBase such that it could be used to
handle non-physical values in any PDE. It is only called in Euler, but
other PDEs should be modified to use it in the future.

As an impact of this change, many constructors needed to be changed.
Rather than adding a single new parameter, I have added a parameters
pointer to PhysicsBase. I think this is a good structural addition.
Constructors in `physics_factory` and `model_factory` have gotten very
complex, and having a parameters object could allow us to simplify these
constructors in the future.

Note that a dummy parameters pointer had to be added to all physics unit
tests to be compatible with the changes to constructors.

CTEST passes as before.

---------

Co-authored-by: Julien Brillon <43619715+jbrillon@users.noreply.github.com>