Adding spanwise divisions parameter as part of the parameter input for naca0012 periodic case.
The purpose of this spanwise refinement is to avoid the failure that occurs when periodic cells are assigned to different processors.

Input meshes must be 1 cell spanwise, which is then refined using deal.ii after the processors are assigned. This ensures that each processor is responsible for the entire width of its segment.