Clean-up DG (#241)
The goal of this PR is to clean up some of the DG code.

1. Separate the DGBase and DGBaseState into different files
2. Introduce the LocalSolution class. It is now used in WeakDG, it still
needs to be adopted elsewhere.
3. Delete unused code, including many of the `assemble_*_explicit()`
functions. Note that `assemble_volume_explicit` still exists. It's
because it is somehow responsible for evaluating max_dt.

The next step in the clean up is to introduce a GlobalSolution class
that can retrieve a LocalSolution. This should be the defacto object to
be held by most of the current data structures that hold a DG for the
sake of accessing the solution. StrongDG should also have changes to
start using LocalSolution.