Fixed exterior reference normal zone issue in strong.  (#250)
This PR fixes/closes the zone issue in strong DG as per issue #249 . The
problem was that previously, on a face, it was assumed that the
reference faces were neighbours, which is the case for grids made from
hypercube and a warping or a single mesh in dealii. In the general case,
this isn't, and now it has been fixed. Note, the reference cell itself
could be rotated and no longer be a square going from 0->1, that case
has not been tested in strong.

![strong_naca0012_solution](https://github.com/dougshidong/PHiLiP/assets/39313421/102c466b-ecc6-4268-90d0-17a7557650ff)

---------

Co-authored-by: Carolyn <c.pethrick@gmail.com>