Fix for output files being overwritten (#255)
When outputting files using the parameter
`output_solution_every_dt_time_intervals` there are instances where
casting to int results in the same `file_number` being used twice. When
the same `file_number` is used twice, the previous output file is
overwritten with the newer output. Examples of these cases are shown
below.


![fileOutputBug_1](https://github.com/dougshidong/PHiLiP/assets/64045004/1b6bff2c-71de-4c70-b08b-6f6788d429bb)

![fileOutputBug_2](https://github.com/dougshidong/PHiLiP/assets/64045004/95bbdbbd-880c-4656-af54-c98cbce841d3)

To fix this I changed the type for `file_number` to double so that the
round function can be used. With this fix, output files are no longer
being overwritten.