Runge Kutta Base Class + Reduced Order Runge Kutta (#267)
For the addition of a standardised Runge-Kutta class and the addition of
the Runge-Kutta class for the Galerkin Method.

## Major Changes
- RungeKuttaBase
- **New** class that contains the base logic of Runge-Kutta methods with
a list of new virtual functions to preform the logic of Runge-Kutta
methods
		- calculate_stage_solution
		- calculate_stage_derivative
		- sum_stages
		- apply_limiter
		- adjust_time_step
	- Parent class is ODESolverBase
- RungeKuttaODESolver
- **Modified** class that will preform the standard Runge-Kutta Method
with the RungeKuttaBase as the parent class
- LowStorageRungeKuttaODESolver
- **Modified** class that will preform the Low Storage Runge Kutta
Methodwith the RungeKuttaBase as the parent class
- GalerkinROMRungeKuttaODESolver
- **New** class that preforms the Galerkin Runge Kutta Method with the
RungeKuttaBase as the parent class
- This new class does **not** have the functionality for limiters and
Relaxation Runge Kutta
- FlowSolver
- Added a time dependent OnlinePOD to flow solver to store snapshots if
needed.
- Requires adding **one** if statement every timestep, and **one** if
statement at the end of the simulation
- Unsteady Reduced order testing
	- Added a new test that will test the following additions
		- GalerkinPODRKSolver Consistency
		- Threshold Function
		- Flowsolver POD
	- Estimated time is ~125 seconds on 8 cores
	

## Minor Changes
- Added number_modes to reduce the number of modes in a POD Basis if the
number is larger. If a threshold and a number of modes are given, only
the number of modes will resolve. Add the following to the parameter
file to use.
	- `set number_modes = int`
- Added `singular_value_threshold` to reduce the number of modes in a
POD Basis based on the cumulative sum of the singular values. Add the
following to the parameter file to use, the double should be below `1.0`
       - `set singular_value_threshold = double #e.g. 0.999` 
- Spelling fixes to numerous files
- Bug fix for Issue #266 
- Solution is to set a constant seed for the random functions and having
the random flag set to true when using random matrices in the optimality
testing function.
- Bug fix for Issue #271
- Added the logic check that Julian requested in
`NumericalFluxFactory::create_convective_numerical_flux`

## Why not the Petrov-Galerkin?
The Petrov Galerkin Method needs to be updated at every timestep as the
LHS is based on the current Jacobian of the residual.
$$\textbf{V}^T\frac{\partial \textbf{R}^{(k)T}}{\partial
\textbf{w}}\textbf{M}\textbf{V}$$
While the Galerkin method is consistent, it is unsure if this method is
consistent with timestepping. From Carlberg 2017, it is very obvious
that the Galerkin method works in the expected way. However, the
Petrov-Galerkin method is complicated and has not been studied to see if
it works in the expected manner.

If it is as simple as changing the LHS, then a simple class addition can
be made to include the LSPG.

## How to Use
The Galerkin Runge Kutta Method is used in the same way as any other
Runge Kutta method with a POD object. Change the following methods in
the parameters file of your choosing.

```
subsection ODE solver
	...
	set ode_solver_type = pod_galerkin_rk_solver
	...
end

subsection reduced order
	...
	set path_to_search = "/path/to/search"
	...
end

```
## Citations
Carlberg, K., Barone, M., & Antil, H. (2017). Galerkin v. least-squares
Petrov–Galerkin projection in nonlinear model reduction. _Journal of
Computational Physics_, _330_, 693–734.
https://doi.org/10.1016/j.jcp.2016.10.033

---------

Co-authored-by: Carolyn <c.pethrick@gmail.com>
Co-authored-by: PranshulThakur <pranshul777@gmail.com>
Co-authored-by: Shruthi Srinivasan <ssaishruthi98@gmail.com>