Clean-up of RK tableau section (#276)
This PR builds from #267 and unifies the RK tableau classes to have a
common base.

There are also some changes to RRK, notably changing from an ODE solver
type (`rrk_explicit_solver`) to a bool (`use_relaxation_runge_kutta`).
Backwards compatibility has been verified. This change is intended to
allow RRK to be used for a more general set of ODE solvers, which may be
added to the code in the future.

---------

Co-authored-by: Tyson Ofstie <tysonofstie@gmail.com>