Remove Raw Pointers (#283)
There are a few raw pointers in the existing code base, as shown by
#281. This PR will remove them and replace them by `std::unique_ptr`
where applicable. The following files are changed by this PR:
- hyper_reduction_comparison.cpp
- hyper_reduced_petrov_galerkin_ode_solver.cpp
- assemble_ECSW_jacobian.cpp
- assemble_ECSW_residual.cpp
- hrom_test_location.cpp
- HROM_error_post_sampling.cpp

The following files are not changed as there were not designed by the
group
- halton.cpp

The choice of a unique pointer was made to ensure that there is no
additional cost to the construction of the pointers, but while also
ensuring that the resulting type is a smart pointer.

I also noticed some typos in the ctest labels and have updated a few in
the HROM tests.