Added new animations to the README file. (#313)
Since we have some fun new test cases, I put together a PR to update the
README.

These test cases are a good demonstration of cases we are able to run
with PHiLiP.

The DMR and Astrophysical Jet are implemented within the code, but as of
now, the Daru Tenaud is on a separate branch. I can merge that test case
as a part of a future PR after Julien's PR goes through.