Fix issue 314 (#316)
I did some restructuring to allow 1D_BURGERS_OOA_ESFR_RANGE_LONG to pass
using the same test script as other ESFR_RANGE tests.

I noticed while updating the test that the previous end time was not
long enough for the error to drop - it was displaying essentially
interpolation error. The order now drops as expected. Note that we have
observed Burgers' equation losing an order at a lower c than c+, hence I
changed a condition in the test file to allow it to pass.

Closes #314.